添加json的encode
