--- interface/JConfigParser/style/base/base.h
@@ -1,20 +1,0 @@
-#ifndef JCONFIG_STYLE_BASE_H
-#define JCONFIG_STYLE_BASE_H
-
-namespace Joger
-{
-    namespace ConfigParser
-    {
-        class CoderBase
-        {
-
-        };
-
-        class FormaterBase
-        {
-            
-        };
-    }
-}
-
-#endif
